chore: format code

+++ src/interfaces/Interfaces.sol
@@ -6,17 +6,21 @@ interface IERC20 {
 
     function approve(address spender, uint256 amount) external returns (bool);
 
-    function transferFrom(address from, address to, uint256 amount) external returns (bool);
+    function transferFrom(
+        address from,
+        address to,
+        uint256 amount
+    ) external returns (bool);
 
     function decimals() external view returns (uint8);
 }
 
 interface IJewelToken {
     function totalBalanceOf(address _holder) external view returns (uint256);
-    
+
     function transferAll(address _to) external;
 
-    function lockOf(address _holder) external view returns(uint);
+    function lockOf(address _holder) external view returns (uint256);
 }
 
 interface ILockedJewelOffer {
